refactor: Game을 RacingCar에 병합 후 RacingCarGame으로 이름 변경
